chore: Ajustes en módulo de gestión de personas y familias
- Ajuste de las entidades `Personas` y `Familias` con sus atributos y
relaciones.
- Mejoras en los servicios `PersonaService` y `FamiliaService` y sus
implementaciones.
- Actualización de los repositorios DAO `IPersonaDao` y `IFamiliaRepo`.
- Ajustes en controladores MVC para manejo de personas
(`PersonasController`, `PersonasRegistrarEditarController`,
`PersonasBuscarController`).
- Modificación de formularios de presentación y validación con
`PersonaForm` y `PersonasBuscarForm`.
- Correcciones y mejoras en las vistas Thymeleaf: agregar, editar,
buscar personas y gestión de familias.
- Validaciones agregadas para campos obligatorios (dni, nombre,
apellido, fecha de nacimiento).
- Validación para evitar duplicados en DNI al registrar personas.
- Ajustes en diseño responsivo con Bootstrap y MDBootstrap.
- Inclusión de script en MySQL para generar 100 registros de personas de
prueba para testing.
- Manejo de excepciones personalizadas con `Excepcion` para mayor
robustez.

Estos ajustes mejoran la integridad de datos y la experiencia de usuario
en la gestión del módulo ONG.

+++ src/main/java/com/TP/ong/ModuloFamilia/AccesoDAO/IFamiliaRepo.java
@@ -21,5 +21,5 @@ public interface IFamiliaRepo extends JpaRepository<Familias, Long> {
     
     
     @Query("SELECT MAX(f.numeroFamilia) FROM Familias f")
-    Integer findMaxNumeroFamilia();
+    Long findMaxNumeroFamilia();
 }
